optim: factorize i2c_master_init() and i2c_master_delete()
