+++ main/INA260.c
@@ -1,6 +1,5 @@
 #include <stdio.h>
 #include "esp_log.h"
-#include "driver/i2c.h"
 
 #include "INA260.h"
