+++ main/INA260.h
@@ -1,7 +1,7 @@
 #ifndef INA260_H_
 #define INA260_H_
 
-#include "i2c_params.h"
+#include "i2c_master.h"
 
 #define INA260_I2CADDR_DEFAULT 0x40 ///< INA260 default i2c address
 #define INA260_REG_CONFIG 0x00      ///< Configuration register
