+++ main/i2c_master.c
@@ -1,0 +1,36 @@
+#include "i2c_master.h"
+
+#define I2C_MASTER_SCL_IO 48        /*!< GPIO number used for I2C master clock */
+#define I2C_MASTER_SDA_IO 47        /*!< GPIO number used for I2C master data  */
+#define I2C_MASTER_FREQ_HZ 400000   /*!< I2C master clock frequency */
+#define I2C_MASTER_TX_BUF_DISABLE 0 /*!< I2C master doesn't need buffer */
+#define I2C_MASTER_RX_BUF_DISABLE 0 /*!< I2C master doesn't need buffer */
+
+/**
+ * @brief i2c master initialization
+ */
+esp_err_t i2c_master_init(void)
+{
+    int i2c_master_port = I2C_MASTER_NUM;
+
+    i2c_config_t conf = {
+        .mode = I2C_MODE_MASTER,
+        .sda_io_num = I2C_MASTER_SDA_IO,
+        .scl_io_num = I2C_MASTER_SCL_IO,
+        .sda_pullup_en = GPIO_PULLUP_ENABLE,
+        .scl_pullup_en = GPIO_PULLUP_ENABLE,
+        .master.clk_speed = I2C_MASTER_FREQ_HZ,
+    };
+
+    i2c_param_config(i2c_master_port, &conf);
+
+    return i2c_driver_install(i2c_master_port, conf.mode, I2C_MASTER_RX_BUF_DISABLE, I2C_MASTER_TX_BUF_DISABLE, 0);
+}
+
+/**
+ * @brief i2c master delete
+ */
+esp_err_t i2c_master_delete(void)
+{
+    return i2c_driver_delete(I2C_MASTER_NUM);
+}
