+++ main/i2c_master.h
@@ -1,0 +1,12 @@
+#ifndef I2C_MASTER_H_
+#define I2C_MASTER_H_
+
+#include "driver/i2c.h"
+
+#define I2C_MASTER_NUM 0            /*!< I2C master i2c port number, the number of i2c peripheral interfaces available will depend on the chip */
+#define I2C_MASTER_TIMEOUT_MS 1000
+
+esp_err_t i2c_master_init(void);
+esp_err_t i2c_master_delete(void);
+
+#endif /* I2C_MASTER_H_ */
