--- main/i2c_params.h
@@ -1,12 +1,0 @@
-#ifndef I2C_PARAMS_H_
-#define I2C_PARAMS_H_
-
-#define I2C_MASTER_SCL_IO 48        /*!< GPIO number used for I2C master clock */
-#define I2C_MASTER_SDA_IO 47        /*!< GPIO number used for I2C master data  */
-#define I2C_MASTER_NUM 0            /*!< I2C master i2c port number, the number of i2c peripheral interfaces available will depend on the chip */
-#define I2C_MASTER_FREQ_HZ 400000   /*!< I2C master clock frequency */
-#define I2C_MASTER_TX_BUF_DISABLE 0 /*!< I2C master doesn't need buffer */
-#define I2C_MASTER_RX_BUF_DISABLE 0 /*!< I2C master doesn't need buffer */
-#define I2C_MASTER_TIMEOUT_MS 1000
-
-#endif /* I2C_PARAMS_H_ */
