+++ main/self_test/self_test.c
@@ -1,8 +1,8 @@
+#include "i2c_master.h"
 #include "DS4432U.h"
 #include "EMC2101.h"
 #include "INA260.h"
 #include "adc.h"
-#include "driver/i2c.h"
 #include "esp_log.h"
 #include "global_state.h"
 #include "nvs_config.h"
