+++ main/tasks/asic_task.c
@@ -5,7 +5,8 @@
 #include <string.h>
 #include "esp_log.h"
 
-#include "driver/i2c.h"
+#include "freertos/FreeRTOS.h"
+#include "freertos/task.h"
 
 static const char *TAG = "ASIC_task";
